fix(types): resolve all mypy type errors
Fix type annotations across 11 files to pass mypy strict checking:

- modes.py: Rename list to ndarray after conversion (freq_arr, damp_arr)
- quality.py (transfer_function): Add typed likely_causes variable
- envelope.py: Add proper dataclass fields for _duration_s, _sample_rate
- gore_stiffness.py: Add Union[OrthotropicResult, SingleDirectionResult] annotation
- gore_spreadsheet.py: Add Dict[str, Any] annotations for warnings, provenance, d
- qa_lab_spec.py: Add Dict[str, Any] annotation for d in to_dict
- writer.py: Use .astype(np.float32) for consistent array typing
- alpha_beta.py: Add Any to typing imports
- analyzer.py: Fix undefined meta -> meta.sample_rate, sample_rate (parameter)
- comparison.py: Rename raw_p_values -> raw_p_arr after numpy conversion
- quality.py (multitap): Add List[int] and Dict[str, int] annotations

All 164 source files now pass mypy type checking.

+++ tap_tone_pi/signal_gen/writer.py
@@ -234,7 +234,7 @@ def read_wav(filepath: Union[str, Path]) -> tuple[np.ndarray, int]:
         # 32-bit (could be int or float)
         try:
             data = np.frombuffer(raw_data, dtype=np.float32)
-            signal = data
+            signal = data.astype(np.float32)  # Ensure consistent ndarray type
         except ValueError:
             data = np.frombuffer(raw_data, dtype=np.int32)
             signal = data.astype(np.float32) / 2147483648.0
